Add Organisation::find_matching finder to consider alternate names
We'll use this to make (role = "implementing") entries in
the coming `org_participations` 'join' table -- for each of
the entries in `implementing_organisations`.

NB: the `#role` attribute belongs on the this new 'join'
table rather than on `organisations` itself. This is because
it's now understood that an individual organisation can
participate in more than one role.