Add a warning to the actual upload page
Some users have duplicated their actual spend data during reporting due
to us not being clear that the upload appends to the data already in the
application as opposed to replacing it.

As a first step towards fixing this issue, we add a strong warning to
users that the upload will append data.

The Linked guidance should be updated along with this change.