Build implementing org participations when activity is new
For unknown reasons it's not possible to build a new
`OrgParticiption` join entity on an unsaved `Activity`. i.e.

  Activity.new#implementing_organisations << implementing_org

does not set `OrgParticipation#activity_id` when you come to save the
new activity. `Activity` then returns a validation error:

  -> Implementing org participations is invalid

This is related to the presence of the condition filtering by
`OrgParticipation#role` on the Activity `has_many :implementing_organisations`
method. Without this condition, the

  Activity#implementing_organisations << org

method works as expected on a new Activity.

NB: this is issue only affect new (unsaved) instances of an Activity.

The workaround is to set the OrgParticipation join entity manually,
which is much less elegant!