DP adds and removes OrgParticipations for implementing orgs thru UI
Rather than adding / editing individual (often duplicated)
`ImplementingOrganisation` entities, a DP now adds or
removes associations to Organisations using the new the
'join'entity `OrgParticipation`.

We show a dropdown list of Organisations of the following
organisation types:

[15 21 24 71 80 90]

These are the distinct `ImplementingOrganisation#organisation_type`
values which have been migrated.

There are 516 of these organisations (`Organisation.integrating.count`)
which makes for a long list, but acceptable for a first cut,
I think?

In a later PR we will drop the `implementing_organisations`
table, once we are sure we've migrated all of its rows to
the new `org_participations` table.