Upgrade standard to fix parser warning
This is the result of pinning parser to ~> 3.0.1.0 (or a 2.7.3
compatible) and `bundle update standard`.

We were on a very old 0.1.8 standard and we can only get to 1.5.0
without warnings by moving parser on to the version that handles 2.7.3.

A simpler thing would be to bump the .ruby-version to 2.7.5 but since
this requires lengthy installs on all incoming dev boxes, we can put
that off for the moment.