Rescue errors and log to Papertrail and Rollbar
The background Job does not re-raise any errors as we don't
want it to retry. We assume that any errors will re-occur
and are due to bad data, a bug in the report building or
some misconfig. The Job rescues any errors and records them
as exceptions in Rollbar for debugging and remediation.