Bundle update with unpinned mini_racer (0.6.2)
It was surprisingly complicated to get an unpinned mini_racer.
Mini_racer requires compilation, and this was failing on version
0.5.0 or higher.

https://github.com/rubyjs/mini_racer/issues/220 appears relevant, and
gives hints about updating bundler and rubygems (we also updated the
hard-coded rake in the Dockerfile for good measure).

After recreating the Gemfile.lock from scratch (because bundler will
automatically install and run the version of bundler from the Gemfile.lock
file) it still didn't work, but told us to run `bundle lock --add-platform x86_64-linux`
which made it successfully work.

This is probably of use to other projects struggling to upgrade mini_racer.