Return timestamped filename from Uploader to caller
Rather than setting the filename of the download at S3 to be
`export-file-{datestamp}.csv`, the S3Uploader is now initialised
with the original filename as well as the `file` to be
uploaded. It can then set the filename of the uploaded file
flexibly and more helpfully as:

 `{original_name}-{timestamp}.{original suffix}`

The uploader now returns a lightweight object with 2
properties:

- `url`
- `timestamped_filename`

for inclusion in the email sent by the `DownloadLinkMailer`.

This will make it easier for the recipient to identify the
download on their local filesystem and give them greater
confidence in opening a downloaded file.