Drop the legacy 'implementing_organisations' table, no longer used
This removes the old "implementing_organisations" table
which is no longer used. "Implementing" is now a `role`
on `OrgParticipation`.

At present this join model is only linking orgs with the "Implementing" role
to their "Activity" but we plan to use this for all type of Organisation, e.g.

```
Activity#has_one
  :delivery_partner
  through: :extending_org_participation

Activity#has_one
 :extending_org_participations,
 -> { where("org_participations.role = 'Extending'") },
 class_name: "OrgParticipation"
```

In this way there will be one table of Organisations and one table of
`OrgParticipations`, with `OrgParticipation#role` distinguishing the types
of participation defined by IATI:

1. *Funding*: The government or organisation which
  provides funds to the activity.

2. *Accountable*: An organisation responsible for
   oversight of the activity and its outcomes

3. *Extending*: An organisation that manages the budget
  and direction of an activity on behalf of the funding
  organisation

4. *Implementing*: The organisation that physically carries
  out the activity or intervention.

See these PRs for the full refactoring of "implementing
organisations":

- https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/pull/1496
- https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/pull/1500
- https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/pull/1502