Strip control characters from some Activity fields
For fields which are of type 'string' or 'text' (and not an array
of those), before we validate we delete almost all control characters to
avoid XML documents we generate being malformed due to containing
unescaped control characters.

(We don't remove tab, carriage return or line-feed as they're
commonly found within normal text, and are permitted as-is within XML
documents. https://www.w3.org/International/questions/qa-controls#support )