Add MacOS platform to Gemfile.lock
Since upgrading Bundler to 2.3.5 and Rubygems to 3.3.5, `bundle install`
will now compile gems with native components rather than auto-selecting
a precompiled version. In some cases those compilations take so long to
be indistinguishable from a hang.

When we add MacOS variants with

```
bundle lock --add-platform x86_64-darwin
```

then precompiled versions are selected and used in dev, preventing hangs
on `bundle install` when compiling nokogiri, sassc et al

Note that I've not been able to find the behaviour change that causes
this; the bundler gem release notes stop at 2.1.4, after which bundler
started shipping with ruby. We were on 2.1.4 before we leapt to 2.3.5
apropos of the mini_racer issues in #1542.