Don't include fields for types of date not present
The IATI XML must include at least one of planned and actual start dates, planned and actual end dates.

However, if the `planned_start_date` was missing, the XML would contain an invalid `activity-date` field with an empty `iso-date`.

We now check for the presence of each date field before including it.

The model validation on the `dates_step` is responsible for ensuring that at least one of each types of start and end dates (planned / actual) are included.