RED: Remove auth0 routes/controller/view
They aren't accessible or useful right now, even as context. We
acknowledge that there are many associated environmental variables and
terraform provisionings and so on. We'll open a separate card for these,
otherwise the PR will become too large to reason about.

Also remove an auth0-dependent view.
While we wanted to remove it later in a sweep-up operation, we have a
missing-keys spec that fails while it is here.

Co-authored-by: Cristina Rotaru <cristina@dxw.com>