RED: Fix all the controller specs
This is about expunging all references to `logged_in_using_omniauth?`.
This is now redundant as it's sufficient to stub `current_user` under
Devise.

We've added an include of Devise::Test::ControllerHelpers in
rails_helper for controller specs only such that we avoid the
Devise::MissingWarden error in those specs.

Co-authored-by: Dragon <david.mckee@dxw.com>