Validate against IATI schema using nokogiri
When you download an organisation XML export, we automatically validate the XML
against the IATI schema, and fail hard if we fail the validation.

The failing URL and the errors that we caught are available as an uncaught
exception in Rollbar etc.

You can get to the failing XML by appending &skip_validation=1 to the failing URL.
(This is also available in the error message.)

We tested our validator against all the IATI example data, but this takes a few
seconds so we're not running them by default, or including them as a part of the repo.
Running rspec with `--tag full_iati_test` will download the data from github and
run them.

We're concerned about the number of failures that occur: we might want to audit how
many XML documents fail, and why.