BEIS can invite users via Devise
We make CreateUser responsible for validating the @user as
passed, which means it needs to be able to accept a nil organisation,
so make UsersController#organisation able to return nil so as not to
fall over when one is not present.

We set a UUID as the user's password such that the model still
validates, accepting that it'll never be read or used.

We do use a Devise protected method, set_reset_password_token, in
UserMailer#welcome; this sets the password reset token on the User
model. We think we're ok with this; the Devise way is to
send_password_reset_instructions and we need to alter that slightly.

Co-authored-by: CristinaRO <cristina@dxw.com>