Tweak codelist names to distinguish Govt sub-codes
Rather than showing duplicate names for the sub-codes of
Donor and Recipent Govt:

11000: Donor Government
11001: Central Government
11002: Local Government
11003: Public corporations
11004: Other public entities in donor country
12000: Recipient Government
12001: Central Government
12002: Local Government
12003: Public corporations
12004: Other public entities in recipient country
...

we now display the hierarchy in the select list labels, by
editing the names of the codes in the "codelist"
configuration file:

11000: Donor Government
11001: Donor Government > Central Government
11002: Donor Government > Local Government
11003: Donor Government > Public corporations
11004: Donor Government > Other public entities in donor country
12000: Recipient Government
12001: Recipient Government > Central Government
12002: Recipient Government > Local Government
12003: Recipient Government > Public corporations
12004: Recipient Government > Other public entities in recipient country
...

An alternative would be to group the select list options
e.g.

*Donor Govt*
11000: Donor Government
11001: Central Government
11002: Local Government
11003: Public corporations
11004: Other public entities in donor country

*Recipient Govt*
12000: Recipient Government
12001: Central Government
12002: Local Government
12003: Public corporations
12004: Other public entities in recipient country

But this would add complexity which is perhaps not needed.

Ultimately this "channel of delivery" step is expected to
become redundant as each "implementing organisation" should
have an assigned channel of delivery code. Once it's
mandatory to associate at least one implementing organisation
with an activity, this step can be removed and an activity's
channel of delivery code(s) can be derived through its
implementing organisations.