Rewrote rake validation task activities:invalid
We replaced the existing `rake invalid_activities` task with one
which reports the error messages for each failing validation and
renamed it to `rake activities:invalid` in line with other namespaced
tasks.

The old `invalid_activities` task only applied to completed activities,
which wasn't mentioned in the documentation. The new one covers all
failures.