Delete home_page sign in spec
The behaviour has changed such that the #authenticate! method no longer
adequately represents what we want to achieve from the feature. We could
log in via Capybara to make this spec do what it used to but since it is
effectively a duplicate of something in user_can_sign_in_spec we delete
it here.

Co-authored-by: Cristina Rotaru <cristina@dxw.com>