Set a generous 5 minutes OTP expiry
This allows for variation in SMS delivery times.

The only configurable parameter via the devise-two-factor gem is
`otp_allowed_drift`. We don’t have access to the ROTP gem’s `interval`,
but we can set the `drift`. Technically this means the code is also
valid 5 minutes before we send it!

Co-authored-by: rgarner <rgarner@zephyros-systems.co.uk>