Partially implement 2-phase 2FA
Override Devise::SessionsController such that we can detect when users
are set up for 2FA and show a verification step following the
email/password login step.

Lots is taken from step 4 of
https://www.jamesridgway.co.uk/implementing-a-two-step-otp-u2f-login-workflow-with-rails-and-devise/

Secrets are generated on a per-user basis in User#ensure_otp_secret! and
that secret is used for that user only to generate OTPs.

This commit has no means of communicating that OTP to users, so at the
time of this commit you can only log in by finding the `User` in a rails
console, then asking `user.current_otp` and pasting the result into the
verification step.

Co-authored-by: Cristina Rotaru <cristina@dxw.com>