Confirm a user's mobile number
When a user has no mobile_number set up, prompt them to enter one by
adding yet another view (mobile_number.html.haml) to the already pretty
heavy Users::SessionsController. We'd like to clean this up, not least
because we need to generate a link to edit the user's mobile number in
the case that it was the wrong one and they never received any messages.
At present this is difficult because every 2FA concept is going through
a before_action of Users::SessionsController#create.

Co-authored-by: Cristina Rotaru <cristina@dxw.com>