Allow users with unconfirmed numbers to edit them
A user’s mobile number is considered “confirmed” after their first
login. Until such time as this confirmation happens we cannot assume
that we have their number on file correctly; show a link to edit the
number as long as they have never logged in.

This is still packing the `users/session` controller somewhat.

Co-authored-by: Cristina Rotaru <cristina@dxw.com>