Require OTP for login by default
`otp_required_for_login` is too entangled with the inner workings of the
gem to remove it outright, but we can make it true by default for all
users.

We’ve had to default the user factory to *not* require OTP for login, in
order to not break all the specs.

Co-authored-by: rgarner <rgarner@zephyros-systems.co.uk>