Add security requirements to the user model
The gem we've chosen to use for password validation also requires
an `EmailValidator` to exist.

The existing Auth0 password requirements are stricter than the BEIS IT
policy of minimum 9 characters, so we went with the higher minimum of
15 characters that users will already be used to.

We’ve modified the default error messages that devise-security comes
with so they can better serve the way we display them on the front-end.

Co-authored-by: rgarner <rgarner@zephyros-systems.co.uk>