Add bulk upload of Refunds
Our chosen method is to bolt this functionality onto the existing Actuals
upload, renaming the Value column to Actual Value and adding a new Refund Value
column.

Exactly one of these values must be filled in for an upload to occur.

Create Actual now takes an optional parameter which allows it to generate a
Refund instead, but is otherwise unchanged.

Import Actual contains logic for determining whether a row is a Refund or
Actual, based on the presence of the two types of value.