Replace i18n references with actual text
We've been finding it quite hard to understand what a test is actually trying to do due to the indirection created by using the translation.

While this may make the tests a bit more brittle, the cost of having to fix a spec if some copy changes is considerably smaller than a developer struggling to understand what a spec is testing.