On password change, redirect to login
This fixes the bug where a Forgotten Password link effectively bypassed
the need for 2FA. Because of the way Devise works by default, we need to
immediately sign the user out and then redirect them to log in with an
appropriate message.

It is much harder to drop them at the right point in the OTP flow to
take into account mobile number confirmation and so on; this is mildly
inconvenient infrequently but saves us from having to hook into a
process not designed for hooking into.

Co-Authored-By: Russell Garner <rgarner@zephyros-systems.co.uk>