On password reset, redirect to login
This fixes the bug where a Forgotten Password link effectively bypassed
the need for 2FA. By default, Devise automatically signs in users after
they’ve reset their password. Setting that option to false changes this
behaviour, and users are redirected to the log in page.

We’ve also modified the text that Devise displays as part of this flow
to make it clearer what the user needs to do.

It would be much harder to drop the user at the right point in the OTP
flow to take into account mobile number confirmation and so on; this is
mildly inconvenient infrequently but saves us from having to hook into a
process not designed for hooking into.

Co-Authored-By: Russell Garner <rgarner@zephyros-systems.co.uk>