Catch GOV.UK mobile phone validation errors
The ruby Notify client doesn’t perform any validation on mobile numbers.
The first time we know something went wrong is when trying to send the
SMS.

We rescue Bad Request errors, and any that have a mobile number
validation error in them we treat as an ActiveModel error. Anything else
we don’t know how to handle, so if the string doesn’t match we re-raise
it explicitly.

We are aware that we have tied our implementation to the strings coming
from GOV.UK Notify API.

Co-authored-by: Russell Garner <rgarner@zephyros-systems.co.uk>