Ensure CSV export has status displayed with numeric code first
e.g. 12345: Kittens and other felines

We've updated just the CSV export to change how this data is presented here, as
well as earlier updating the website editing interface.

This is a slightly ugly wart: the ActivityCSVPresenter now overrides the
behavior of the ActivityPresenter in a quite surprising way.

It's possible that it would have been better to just change the behaviour
everywhere to include the code, but exploring that path can be left for another
time.