Disentangle activity-comments associations
Having `activity.comments` be a readonly association breaks the principle
of least surprise, and leads to confusion when trying to create and
access an activity’s direct comments, as opposed to comments left on the
activity and on all its direct child transactions (actuals, refunds, and
adjustments).

We are therefore creating a separate, hopefully more clearly named
association for comments belonging to the activity and its child
transactions, and replacing the use of `activity.comments` with this new
association wherever relevant.