BEIS users cannot create comments on actuals
It is already the case that BEIS users cannot create comments on
activities.

Rather than allow comments to not belong to a report, it seems to make
more sense to disallow BEIS users to add comments to actuals. (Level B
AKA programmes do not require a report to exist in order to have actuals.)