Fix Bullet's wagon for Report comments
Comments are polymorphic, and we get conflicting advice from Bullet to
both remove AND add .includes in different spec contexts; some for
Activity, some for Transaction descendants.

Adding to the safelist in environments/test.rb causes whack-a-mole
errors. This seems to be the safest way of indicating that we're cutting
the Gordian knot in a localised fashion.

Co-authored-by: Cristina Rotaru <cristina@dxw.com>