Delete duplicate organisations
We deleted a number of organisations at BEIS' request. They had no
matched_efforts, reports, activities, users, or budgets referring to them
but did have org_participations. We deleted those.

There is an unused SQL query in the migration which we used to verify the above
and the deletion afterwards, left in the hopes that it might be useful.

A number of other organisations were already deleted.

See Zendesk ticket 16066.