Ensure table cells are not empty for screenreaders
Accessibility standards require table cells to not be empty, so that
screenreader users get the relevant information - including the
information that a cell is empty because there is no data, not because
of a lack of thought.

On some tables we chose to make the “Actions” header visible, while on
others we decided that visual users don’t need it. We should still
provide this header for screenreader users.