Remove invalid “aria-controls” attribute
The `controls` attribute, for an element with the `tab` role, is meant
to reference the ID of the tab panel that the tab element controls.

This doesn’t seem to be needed when the panel being controlled is on a
different page, thus not in the current page structure. (For an example
of tab panels being on the same page, see the Reports index page, where
Current and Approved reports are rendered on the same page.)