"Reset password" links expire after 24 hours
We had extended the validity of "Reset password" links to two weeks to enable
users recieving the password reset link that we'd sent them sufficient time to
log in with the link provided.

The 31st of March is the fourteen-day anniversary of us making that change --
the reset links will become invalid soon, so now is a good time to change this
setting back to something more reasonable.

There was some discussion about the most appropriate amount of time: 24 hours
is an arbitarily-chosen compromise.