WIP: starting point for terraforming s3-export-download-bucket
The bucket creation/config which we need to automate is
something like:

1. Create bucket
----------------

We'd like buckets "named" in the
form `s3-export-download-bucket-staging`, e.g. producing a
`VCAP_SERVICE` env like:

```
VCAP_SERVICES='{
   "aws-s3-bucket":[
      {
         "name":"s3-export-download-bucket-staging",
         "credentials":{
            "bucket_name":"paas-s3-broker-prod-lon-id",
            "aws_access_key_id":"KEY_ID",
            "aws_secret_access_key":"SECRET",
            "aws_region":"eu-west-2",
            "deploy_env":""
         }
      }
   ]
}'
```

They need to be public buckets so that users can download
their reports using links which we email to them.

In the `cf` CLI this is
done with a command like:

```
cf create-service aws-s3-bucket default \
  s3-export-download-bucket-staging \
  -c '{"public_bucket":true}'
```

2. Bind the service to the app
------------------------------

I think the service needs to be bound to the worker app as
that's the instance which will be creating the artefact
which will be stored on S3. On `cf` I think this is:

```
cf bind-service beis-roda-staging-worker \
  s3-export-download-bucket-staging \
  -c '{"permissions":"read-write"}'
```

I'm not clear how we go about setting the permissions here
as the GPaaS docks suggest that you "must use the command line
when you want to specify permissions for the bucket" but I'm
think that's only considering their 'manifest' rather than
full-blown terraform?

https://docs.cloud.service.gov.uk/deploying_services/s3/#bind-an-aws-s3-bucket-to-your-app