Remove invalid countries from 57 activities benefitting_countires
Support ticket https://dxw.zendesk.com/agent/tickets/15636 brought it to our
attention that there are 33 countries which have not been eligible during the period
covered by RODA.

57 activities have included these countries in their list of #benefitting_countries.
This is due to the mechanism where when a grouping "region" is selected, all of the
constituent countries are added to the Activity#benefitting_countries list.

In the data migration we:

- print out information on:
  - the 33 countries and affected activities
  - the affected 57 activities
- remove the invalid countries from each of the affected 57 activities'
  #benefitting_countries list

REMOVING THE 33 COUNTRIES THEMSELVES
This will be done in a separate PR/Release as we need the 33
entries in `vendor/data/codelists/BEIS/benefitting_countries.yml`
for the migration.