Generate public rather than pre-signed url for download link
It seems to be impossible to generate a pre-signed GET url
using the GPaaS S3 config, either on a bucket which is
created as "public" or "private".

I've identified two ways of generating a working download
link using Terraform / GPaaS:

1. Generate a public url rather than a pre-signed url
---
This has the disadvantage that the links never expire.
However, I think it's adequate in terms of security as the
url is essentially unguessable e.g.

https://paas-s3-broker-prod-lon-8cc9c2c5-43f7-44bf-a5c1-11c4256feba8.s3.eu-west-2.amazonaws.com/GCRF_spending_breakdown-20220411170906.csv

2. Create a pre-signed url using separate "service key" creds
---
By generating a separate set of AWS credentials which are
explicitly set to enable external access:

```
cf create-service-key \
  beis-roda-staging-s3-export-download-bucket \
  beis-roda-staging-s3-export-download-bucket \
  -c '{"allow_external_access": true}'
```

and using these to generate a pre-signed URL (on a private
bucket).

However, these "service key" credentials are not
readily available as environment variables. I've asked GPaaS
support for some help in understanding how we might access
them in Terraform.

Recommendation
---
I think that (1): the use of public url is a better
(simpler) solution which is adequate in terms of governance
and security.