Describe how names of all implementing organisations are included
There was some confusion in the code review stage about
where existing implementing organisations would be listed in
a downloaded CSV.

It's important to make this data available to Delivery
Partners to minimise the risk of inadvertently
disassociating implementing organisations by not including
existing ones when doing an update.

On investigation we see that the behaviour for *activities* is
different from that of *actuals*:

1. The activity upload template does NOT include data
The **activity** upload template only ever
includes a header row:

https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/blob/1e382c87d9af0d05a81fc9920ce149869808619f/app/controllers/staff/activity_uploads_controller.rb#L19-L26

That's in contrast to the **actuals**
upload template which does include a row for each activity:

https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/blob/1e382c87d9af0d05a81fc9920ce149869808619f/app/controllers/staff/actual_uploads_controller.rb#L19-L30

2. The report download CSV already includes implementing organisations

From the Report > Activities tab, when you "Download Report as CSV"
you receive a list of activities, and there is a column
"Implementing organisations" with a pipe-separated list of
organisation names.

The spec for the `Export::ActivityImplementingOrganisationColumn`
from Nov 2021 describes how multiple implementing organisations
are represented in the form 'Org Name 1|Org Name 2':

https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/blob/1e382c87d9af0d05a81fc9920ce149869808619f/spec/models/export/activity_implementing_organisation_column_spec.rb#L35-L50