Use graduated country as 'benefitting' in 4 historic cases
Since:

https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/commit/fe68bc2463512c73b6f594b9bc1d19890dbb6b13

(Sept 2021), when adding or updating an activity via the bulk
upload mechanism, it's not been possible to associate a "graduated"
country with an activity as an entry in `benefitting_countries`. However,
there are some historic activities (they have `actual_end_dates`)
which are missing a benefitting country which was valid at
the time it was active.

For example, GCRF-AH_C_LGI-2016AH/P009158/1 was active between
01-Nov-2016 and 30-Jun-2017.

At that time Chile was a valid benefitting country. However, the
historic activity was added to RODA last year (15-Jun-2021)
and by that later date Chile was no longer valid for ODA --
it had the status "graduated".

In this migration we add "graduated" countries to the list of
benefitting countries in the case of 4 historic activities.