Use locale string for welcome email subject
We need the email subject to be more flexible than the Notify template
personalisations are allowing it to be: we want to conditionally display
some content, and we also want that optional content to vary depending
on environment name.

Notify templates do not allow empty personalisations, and optional
content cannot be personalised.

Co-authored-by: George <george@dxw.com>