Don't create rows with no export data
The previous implementation was calling `forecast_rows.any?` on some
forecast rows with the following structure:

```
{ 'activity-id'=>[100] }
```

This meant that we were calling `any?` on a hash value which was always
returning `true`, in contrast to the `headers` which would return an
empty array.

This led to the number of headers being out of sync with the number of
rows, causing columns to be out of sync from the `variance` row.

Now we check all the arrays in that hash are empty - if they are, we
don't render the variance row.

For the tests, We've added an entire new spec because the existing setup sets up a forecast, and
there's no easy way we could see to remove it or set it up at a later stage. It's intended
to otherwise be a complete copy of the setup for the with-forecast case.