Fix misspelling of “FSTC” in codelist
FSTC stands for “Free-standing Technical Cooperation”.[1]

The misspelling in the codelist is consistent with the code that checks
for that key, and the activity attribute is correctly spelled. However,
the misspelling is a source of confusion, so we want to fix it.

[1] If we were to use the official acronym, via OECD:
https://stats.oecd.org/glossary/detail.asp?ID=6022
we would be switching all of them to “FTC”. However, I feel that would
be more disruptive than helpful at this point.