Fix spec that was passing by accident
The spec was outside the “as a delivery partner” context, so the report
was not visible on the page because the visitor was not authenticated.
This is not what we intended to test, as authentication is tested
elsewhere.

After moving the test in the DP user scope, the test was still passing
by coincidence, since it was not the approved state that made it not
visible to the user, but the fact that the report was not scoped to the
organisation of the DP, which is already tested above.

After creating an approved report belonging to the user’s organisation,
we find that the test fails, because the current and approved reports
are on the same HTML page, and we control the visibility with tabs.

In order to have a passing test, we need to narrow our assertion to the
current tab.