Add extra info about activity to report CSVs
The extra information consists of the parent programme RODA ID and
title, and the project RODA ID and title where applicable.

Where not applicable (e.g. for a project, AKA level C), the columns are
still added, and their values are left blank.

The `includes` clause was added to optimise the query as per Bullet’s
demands.