Fix typos in partials
There was a typo in the instance var names being passed to the partial.
In addition, there was an error in how we passed the locals from the
`service_owner` view to the `delivery_partner_organisations_table`
partial.

The reason we never noticed is that we weren’t relying on the locals at
all. Instead, the partials were directly using the instance vars (which
are correctly set in the controllers).

I only noticed when I started replacing the names using “delivery
partner” with “partner organisation” and had to take a closer look at
these files.

I’m splitting this fix into its own commit in order to make the commit
with the renaming self-contained and easier to review.