Rename `delivery_partner` role
The values in the database are integers (thanks to the attributes being
`enum`s), so they don’t need a data migration.

Renaming the roles has to happen at the same time as renaming all the
methods, locale strings, partials etc. that use the role names instead
of their numeric values.