Refactor activity upload partials
This DRYs up the activity upload partials that were duplicated in
the original activity bulk upload and the new Level B activity
bulk upload. These are now within an activity_uploads folder in the
shared partials folder. The activities table partial is also used
by the staff/reports/acitivites view (the activities tab when
viewing a report), so this is also updated to use the shared partial

As the two URLs in the upload form partial differ slightly - one
requires the format to be CSV, whereas the other doesn't - the path
helper is sent as a local and invoked using `send` with different
arguments in order to render the appropriate URL

Co-authored-by: George <george@dxw.com>