Reset env var post-spec
The spec `it "loads redis URL to the ENV”` has side effects that can
cause other specs to fail, depending on the order in which they are run.

`ClimateControl` can only reset `VCAP_SERVICES` for us, not the other
env vars that are set during the spec running.

The default value of `VCAP_SERVICES` (set in .env.test) doesn’t have a
value for `REDIS_URL`, so the value being set during this spec sticks,
and overwrites the desired value of `REDIS_URL`, which is set in
.env.test.