Ensure logged out specs start with clean state
Specs that rely on there being no authenticated user in order to pass
should ensure that there are no authenticated users leaking in from
other specs.

We’ve also thought about ensuring that specs that authenticate a user
clean up after themselves, but there are a lot more of those specs than
there are specs relying on no logged in user.

The lack of setup was causing some specs to fail intermittently,
depending on the order of running.