Use Level B policy for bulk activity upload action authorisation
Previously the Level B activity uploads controller was using the organisation
policy to authorize its actions. The previous commit introduced a Level B policy
as part of a move towards a new namespacing approach. This brings the controller
into the remit of the Level B policy, adding a new `activity_upload?` method

Consolidating the two methods (`budget_upload?` and `activity_upload?`) was
considered, but we opted for the more granular approach that has been taken in
other policies, which makes it easier in future to adapt to any changes in who
is allowed to perform specific actions